pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gismo"
version = "0.1.0"
description = "GISMO persistent orchestration core"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "UNLICENSED"}
dependencies = [
    "textual>=0.80",
    "piper-tts>=1.4",
]

[project.scripts]
gismo = "gismo.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gismo*"]
